Fix many small warnings in flake8 (#962)
Fixed the following:
DOC404: The types in the docstring’s Yields section and the return annotation in the signature are not consistent
ANN102: Missing type annotation for cls in classmethod
ANN206: Missing return type annotation for classmethod
VNE001: single letter variable names like 'XXX' are not allowed
DOC105: Argument names match, but type hints do not match
DOC102: Docstring contains more arguments than in function signature
DOC301: init() should not have a docstring; please combine it with the docstring of the class
DCO024: function/ method has one or more arguments described in the docstring which are not arguments of the function/ method.
VNE002: variable name 'XXX' should be clarified
B009: Do not call getattr(x, 'attr'), instead use normal property access: x.attr. Missing a default to getattr will cause an AttributeError to be raised for non-existent properties. There is no additional safety in using getattr if you know the attribute name ahead of time.
DOC001: Potential formatting errors in docstring
DCO053: function/ method that raises one or more exceptions where one or more of the exceptions is not described in the docstring.
DCO023: function/ method has one or more arguments not described in the docstring.
VNE003: variable names that shadow builtins are not allowed
DCO050: function/ method raises one or more exceptions and the docstring does not have a raises section.
DOC501: Function/method has “raise” statements, but the docstring does not have a “Raises” section