Make the burn command work on message reply (#1298)
the `.burn` command now behaves like the `.hug` command, meaning it doesnt require mentioning a user to mark their message as a burn. replying to the message works. it still retains it's original behavior of marking the last message of a user as a burn when it is used as `.burn @user`.